[Modify] 로그인 전/후 로직 분리 & [Refactor] Carousel 컴포넌트 리팩토링

+++ src/components/WishBtn/WishBtn.jsx
@@ -3,7 +3,7 @@ import axios from "axios";
 import { FaHeart, FaRegHeart } from "react-icons/fa";
 import "./WishBtn.css";
 
-const Wishlist = ({ memberId, productId }) => {
+const WishBtn = ({ memberId, productId, isLoggedin, pageType }) => {
   const [isWishlist, setIsWishlist] = useState(false);
 
   useEffect(() => {
@@ -27,32 +27,57 @@ const Wishlist = ({ memberId, productId }) => {
   };
 
   const handleWishSubmit = async () => {
-    try {
-      if (!isWishlist) {
-        // Add to wishlist
-        const response = await axios.post(
-          `http://localhost:8080/api/v1/wishlist/${productId}/${memberId}`
-        );
-        console.log(response.data);
-        setIsWishlist(true);
-      } else {
-        // Remove from wishlist
-        const response = await axios.delete(
-          `http://localhost:8080/api/v1/wishlist/${productId}/${memberId}`
-        );
-        console.log(response.data);
-        setIsWishlist(false);
+    if (isLoggedin === true) {
+      try {
+        if (!isWishlist) {
+          // Add to wishlist
+          const response = await axios.post(
+            `http://localhost:8080/api/v1/wishlist/${productId}/${memberId}`
+          );
+          console.log(response.data);
+          setIsWishlist(true);
+        } else {
+          // Remove from wishlist
+          const response = await axios.delete(
+            `http://localhost:8080/api/v1/wishlist/${productId}/${memberId}`
+          );
+          console.log(response.data);
+          setIsWishlist(false);
+        }
+      } catch (error) {
+        console.error(error.response.data);
       }
-    } catch (error) {
-      console.error(error.response.data);
+    } else {
+      alert("로그인 후 이용 가능해요!");
     }
   };
 
-  return (
-    <span className="heart-btn" onClick={() => handleWishSubmit()}>
-      {!isWishlist ? <FaRegHeart /> : <FaHeart />}
-    </span>
-  );
+  const renderWishButton = () => {
+    if (pageType === "productDetail") {
+      return (
+        <li className="pd-page-heart-btn" onClick={() => handleWishSubmit()}>
+          {!isWishlist ? <FaRegHeart /> : <FaHeart />}
+          <span>찜하기</span>
+        </li>
+      );
+    } else if (pageType === "cart") {
+      return (
+        <span
+          className="cart-page-heart-btn"
+          onClick={() => handleWishSubmit()}
+        >
+          {!isWishlist ? <FaRegHeart /> : <FaHeart />}
+        </span>
+      );
+    }
+  };
+
+  // return (
+  //   <span className="heart-btn" onClick={() => handleWishSubmit()}>
+  //     {!isWishlist ? <FaRegHeart /> : <FaHeart />}
+  //   </span>
+  // );
+  return renderWishButton();
 };
 
-export default Wishlist;
+export default WishBtn;
